// src/wallet/sdk/Wallet.interfaces.ts
export type BooleanDefaultFalse = boolean

/**
 * @typedef {boolean} BooleanDefaultTrue
 * Represents an optional boolean parameter, which defaults to `true` if not provided.
 */
export type BooleanDefaultTrue = boolean

/**
 * @typedef {number} Byte
 * Represents an integer from 0 to 255 (inclusive).
 * @minimum 0
 * @maximum 255
 */
export type Byte = number
// ...
export type ISOTimestampString = string

/**
 * @typedef {string} HexString
 * A string containing only hexadecimal characters (0-9, a-f).
 */
export type HexString = string
// ...
export type PubKeyHex = HexString

/**
 * @typedef {string} Base64String
 * A standard base64 encoded string.
 */
export type Base64String = string

/**
 * @typedef {string} OriginatorDomainNameStringUnder250Bytes
 * Represents the fully qualified domain name (FQDN) of the application that originates the request.
 */
export type OriginatorDomainNameStringUnder250Bytes = string

/**
 * @typedef {string & { minLength: 5, maxLength: 50 }} DescriptionString5to50Bytes
 * A string used for descriptions, with a length between 5 and 50 characters.
 */
export type DescriptionString5to50Bytes = string
// ...
export type ProtocolString5To400Bytes = string

/**
 * @typedef {string & { maxLength: 800 }} KeyIDStringUnder800Bytes
 * Represents a key identifier string, with a maximum length of 800 characters.
 */
export type KeyIDStringUnder800Bytes = string
// ...
export type WalletProtocol = [0 | 1 | 2, ProtocolString5To400Bytes]

export type WalletCounterparty = PubKeyHex | 'self' | 'anyone'
// ...
export interface WalletEncryptionArgs {
  protocolID: WalletProtocol
  keyID: KeyIDStringUnder800Bytes
  counterparty?: WalletCounterparty
  privileged?: BooleanDefaultFalse
  privilegedReason?: DescriptionString5to50Bytes
  seekPermission?: BooleanDefaultTrue
}
// ...
export interface GetPublicKeyArgs extends Partial<WalletEncryptionArgs> {
  identityKey?: true
  forSelf?: BooleanDefaultFalse
}
// ...
export interface RevealCounterpartyKeyLinkageArgs {
  counterparty: PubKeyHex
  verifier: PubKeyHex
  privileged?: BooleanDefaultFalse
  privilegedReason?: DescriptionString5to50Bytes
}
// ...
export interface RevealSpecificKeyLinkageArgs {
  counterparty: WalletCounterparty
  verifier: PubKeyHex
  protocolID: WalletProtocol
  keyID: KeyIDStringUnder800Bytes
  privilegedReason?: DescriptionString5to50Bytes
  privileged?: BooleanDefaultFalse
}

/**
 */
export interface KeyLinkageResult {
  encryptedLinkage: Byte[]
  encryptedLinkageProof: Byte[]
  prover: PubKeyHex
  verifier: PubKeyHex
  counterparty: PubKeyHex
}

/**
 */
export interface RevealCounterpartyKeyLinkageResult extends KeyLinkageResult {
  revelationTime: ISOTimestampString
}

/**
 */
export interface RevealSpecificKeyLinkageResult extends KeyLinkageResult {
  protocolID: WalletProtocol
  keyID: KeyIDStringUnder800Bytes
  proofType: Byte
}

/**
   * @param {Byte[]} plaintext - Array of bytes constituting the plaintext data to be encrypted.
 */
export interface WalletEncryptArgs extends WalletEncryptionArgs {
  plaintext: Byte[]
}

export interface WalletEncryptResult {
  ciphertext: Byte[]
}

/**
   * @param {Byte[]} ciphertext - Encrypted bytes, including the initialization vector, for decryption.
 */
export interface WalletDecryptArgs extends WalletEncryptionArgs {
  ciphertext: Byte[]
}

export interface WalletDecryptResult {
  plaintext: Byte[]
}

/**
   * @param {Byte[]} data - Input data (in bytes) for which the HMAC needs to be created.
 */
export interface CreateHmacArgs extends WalletEncryptionArgs {
  data: Byte[]
}

export interface CreateHmacResult {
  hmac: Byte[]
}

/**
   * @param {Byte[]} data - The input data whose HMAC is to be verified.
   * @param {Byte[]} hmac - Byte array representing the HMAC value to be verified.
 */
export interface VerifyHmacArgs extends WalletEncryptionArgs {
  data: Byte[]
  hmac: Byte[]
}

export interface VerifyHmacResult {
  valid: boolean
}

/**
   * @param {Byte[]} [data] - Data to be signed using the derived private key with ECDSA. Required unless directly signing a hash.
   * @param {Byte[]} [hashToDirectlySign] - Sign a pre-hashed value in situations where data can't or shouldn't be revealed, whether due to its size or for privacy.
 */
export interface CreateSignatureArgs extends WalletEncryptionArgs {
  data?: Byte[]
  hashToDirectlySign?: Byte[]
}

export interface CreateSignatureResult {
  signature: Byte[]
}
// ...
export interface VerifySignatureArgs extends WalletEncryptionArgs {
  data?: Byte[]
  hashToDirectlyVerify?: Byte[]
  signature: Byte[]
  forSelf?: BooleanDefaultFalse
}

export interface VerifySignatureResult {
  valid: true
}